Week2/Day1 - Refactor tail
