+++ tail.js
@@ -1,29 +1,7 @@
-// FUNCTION IMPLEMENTATION
-const assertEqual = function(actual, expected) {
-  if (actual === expected) {
-    console.log(`\uD83D\uDE00 Assertion Passed: ${actual} === ${expected}`);
-  } else {
-    console.log(`\uD83D\uDE1E Assertion Failed: ${actual} !== ${expected}`);
-  }
-};
-
-// TEST CODE
-assertEqual("Lighthouse Labs", "Bootcamp");
-assertEqual(1, 1);
-assertEqual("bla","bla1");
-assertEqual(2,2);
-assertEqual(234,342);
+const assertEqual = require('./assertEqual');
 
 const tail = function(value) {
   return value.slice(1);
 };
 
-arrayTest = ["Hello", "Lighthouse", "Labs"];
-console.log(tail(["Hello", "Lighthouse", "Labs"]));
-console.log(arrayTest);
-
-const result = tail(["Hello", "Lighthouse", "Labs"]);
-assertEqual(result.length, 2); // ensure we get back two elements
-assertEqual(result[0], "Lighthouse"); // ensure first element is "Lighthouse"
-assertEqual(result[1], "Labs"); // ensure second element is "Labs"
-
+module.exports = tail;
