+++ test/tailTest.js
@@ -1,0 +1,12 @@
+const tail = require('../tail');
+const assertEqual = require('../assertEqual');
+
+arrayTest = ["Hello", "Lighthouse", "Labs"];
+//console.log(tail(["Hello", "Lighthouse", "Labs"]));
+//console.log(arrayTest);
+
+const result = tail(["Hello", "Lighthouse", "Labs"]);
+assertEqual(result.length, 2); // ensure we get back two elements
+assertEqual(result[0], "Lighthouse"); // ensure first element is "Lighthouse"
+assertEqual(result[1], "Labs"); // ensure second element is "Labs"
+assertEqual(result, "Labs");
